+++ examples/classification/4-tutorials/plot_crossconformal.py
@@ -108,7 +108,7 @@ plt.show()
 
 kf = KFold(n_splits=5, shuffle=True)
 clfs, mapies, y_preds, y_ps_mapies = {}, {}, {}, {}
-methods = ["label", "aps"]
+methods = ["lac", "aps"]
 alpha = np.arange(0.01, 1, 0.01)
 for method in methods:
     clfs_, mapies_, y_preds_, y_ps_mapies_ = {}, {}, {}, {}
@@ -132,8 +132,8 @@ for method in methods:
 # set and the estimated quantile for ``alpha`` = 0.1.
 
 
-fig, axs = plt.subplots(1, len(mapies["label"]), figsize=(20, 4))
-for i, (key, mapie) in enumerate(mapies["label"].items()):
+fig, axs = plt.subplots(1, len(mapies["lac"]), figsize=(20, 4))
+for i, (key, mapie) in enumerate(mapies["lac"].items()):
     axs[i].set_xlabel("Conformity scores")
     axs[i].hist(mapie.conformity_scores_)
     axs[i].axvline(mapie.quantiles_[9], ls="--", color="k")
@@ -199,12 +199,12 @@ def plot_results(
 
 
 plot_results(
-    mapies["label"],
+    mapies["lac"],
     X_test,
     X_test_distrib,
     y_test_distrib,
     alpha[9],
-    "label"
+    "lac"
 )
 
 plot_results(
@@ -277,7 +277,7 @@ split_widths = np.array(
 )
 
 plot_coverage_width(
-    alpha, split_coverages[0], split_widths[0], "label"
+    alpha, split_coverages[0], split_widths[0], "lac"
 )
 
 plot_coverage_width(
@@ -310,7 +310,7 @@ plot_coverage_width(
 #    the conformity scores from the associated model for a new test point
 #    (as presented in Romano et al. 2020 for the "aps" method)
 #
-# Let's explore the two possibilites with the "label" method using
+# Let's explore the two possibilites with the "lac" method using
 # :class:`~mapie.classification.MapieClassifier`.
 #
 # All we need to do is to provide with the `cv` argument a cross-validation
@@ -338,11 +338,11 @@ kf = KFold(n_splits=5, shuffle=True)
 
 STRATEGIES = {
     "score_cv_mean": (
-        Params(method="label", cv=kf, random_state=42),
+        Params(method="lac", cv=kf, random_state=42),
         ParamsPredict(include_last_label=False, agg_scores="mean")
     ),
     "score_cv_crossval": (
-        Params(method="label", cv=kf, random_state=42),
+        Params(method="lac", cv=kf, random_state=42),
         ParamsPredict(include_last_label=False, agg_scores="crossval")
     ),
     "cum_score_cv_mean": (
@@ -402,7 +402,7 @@ plot_coverage_width(
     alpha,
     [coverages["score_cv_mean"], coverages["score_cv_crossval"]],
     [widths["score_cv_mean"], widths["score_cv_crossval"]],
-    "label",
+    "lac",
     comp="mean"
 )
 
@@ -426,12 +426,12 @@ plot_coverage_width(
 # target coverage between the cross-conformal and split-conformal methods.
 
 violations_df = pd.DataFrame(
-    index=["label", "aps"],
+    index=["lac", "aps"],
     columns=["cv_mean", "cv_crossval", "splits"]
 )
-violations_df.loc["label", "cv_mean"] = violations["score_cv_mean"]
-violations_df.loc["label", "cv_crossval"] = violations["score_cv_crossval"]
-violations_df.loc["label", "splits"] = np.stack(
+violations_df.loc["lac", "cv_mean"] = violations["score_cv_mean"]
+violations_df.loc["lac", "cv_crossval"] = violations["score_cv_crossval"]
+violations_df.loc["lac", "splits"] = np.stack(
     [
         np.abs(cov - (1 - alpha)).mean()
         for cov in split_coverages[0]